# Admitix- Root Folder/Test Cases/admitix_api_test_kit/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Admitix API",
    "description": "Production Ready Management API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/departments/": {
      "get": {
        "tags": [
          "Departments"
        ],
        "summary": "Get Departments Route",
        "operationId": "get_departments_route_departments__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DepartmentRead"
                  },
                  "type": "array",
                  "title": "Response Get Departments Route Departments  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Departments"
        ],
        "summary": "Create Department Route",
        "operationId": "create_department_route_departments__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/departments/{department_id}": {
      "get": {
        "tags": [
          "Departments"
        ],
        "summary": "Get Department By Id Route",
        "operationId": "get_department_by_id_route_departments__department_id__get",
        "parameters": [
          {
            "name": "department_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Department Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Departments"
        ],
        "summary": "Update Department Route",
        "operationId": "update_department_route_departments__department_id__patch",
        "parameters": [
          {
            "name": "department_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Department Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Departments"
        ],
        "summary": "Delete Department Route",
        "operationId": "delete_department_route_departments__department_id__delete",
        "parameters": [
          {
            "name": "department_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Department Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/document-types/": {
      "get": {
        "tags": [
          "Document Types"
        ],
        "summary": "Get Document Types Route",
        "operationId": "get_document_types_route_document_types__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentTypeRead"
                  },
                  "type": "array",
                  "title": "Response Get Document Types Route Document Types  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Document Types"
        ],
        "summary": "Create Document Type Route",
        "operationId": "create_document_type_route_document_types__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentTypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentTypeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/document-types/{document_type_id}": {
      "get": {
        "tags": [
          "Document Types"
        ],
        "summary": "Get Document Type By Id Route",
        "operationId": "get_document_type_by_id_route_document_types__document_type_id__get",
        "parameters": [
          {
            "name": "document_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentTypeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Document Types"
        ],
        "summary": "Update Document Type Route",
        "operationId": "update_document_type_route_document_types__document_type_id__patch",
        "parameters": [
          {
            "name": "document_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentTypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentTypeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Document Types"
        ],
        "summary": "Delete Document Type Route",
        "operationId": "delete_document_type_route_document_types__document_type_id__delete",
        "parameters": [
          {
            "name": "document_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Type Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Documents Route",
        "operationId": "get_documents_route_documents__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentRead"
                  },
                  "type": "array",
                  "title": "Response Get Documents Route Documents  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Create Document Route",
        "operationId": "create_document_route_documents__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document By Id Route",
        "operationId": "get_document_by_id_route_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Documents"
        ],
        "summary": "Update Document Route",
        "operationId": "update_document_route_documents__document_id__patch",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete Document Route",
        "operationId": "delete_document_route_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DepartmentCreate": {
        "properties": {
          "department_name": {
            "type": "string",
            "maxLength": 150,
            "title": "Department Name"
          },
          "hod_staff_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hod Staff Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "boolean",
            "title": "Status",
            "default": true
          },
          "faculty_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faculty Id"
          },
          "institution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institution Id"
          }
        },
        "type": "object",
        "required": [
          "department_name",
          "faculty_id",
          "institution_id"
        ],
        "title": "DepartmentCreate"
      },
      "DepartmentRead": {
        "properties": {
          "department_name": {
            "type": "string",
            "maxLength": 150,
            "title": "Department Name"
          },
          "hod_staff_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hod Staff Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "boolean",
            "title": "Status",
            "default": true
          },
          "department_id": {
            "type": "string",
            "format": "uuid",
            "title": "Department Id"
          },
          "faculty_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faculty Id"
          },
          "institution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institution Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "department_name",
          "department_id",
          "faculty_id",
          "institution_id",
          "created_at"
        ],
        "title": "DepartmentRead"
      },
      "DepartmentUpdate": {
        "properties": {
          "department_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Name"
          },
          "hod_staff_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hod Staff Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "DepartmentUpdate"
      },
      "DocumentCreate": {
        "properties": {
          "file_name": {
            "type": "string",
            "maxLength": 255,
            "title": "File Name"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "document_type_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Type Id"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "file_url",
          "application_id",
          "document_type_id"
        ],
        "title": "DocumentCreate"
      },
      "DocumentRead": {
        "properties": {
          "file_name": {
            "type": "string",
            "maxLength": 255,
            "title": "File Name"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "document_type_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Type Id"
          },
          "verification_status": {
            "$ref": "#/components/schemas/DocumentVerificationStatus"
          },
          "verified_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified By"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "file_url",
          "document_id",
          "application_id",
          "document_type_id",
          "verification_status",
          "verified_by",
          "uploaded_at"
        ],
        "title": "DocumentRead"
      },
      "DocumentTypeCreate": {
        "properties": {
          "document_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Document Name"
          },
          "mandatory": {
            "type": "boolean",
            "title": "Mandatory",
            "default": true
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "document_name"
        ],
        "title": "DocumentTypeCreate",
        "description": "Payload for `POST /document-types`."
      },
      "DocumentTypeRead": {
        "properties": {
          "document_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Document Name"
          },
          "mandatory": {
            "type": "boolean",
            "title": "Mandatory",
            "default": true
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "document_type_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Type Id"
          }
        },
        "type": "object",
        "required": [
          "document_name",
          "document_type_id"
        ],
        "title": "DocumentTypeRead"
      },
      "DocumentTypeUpdate": {
        "properties": {
          "document_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Name"
          },
          "mandatory": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mandatory"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "DocumentTypeUpdate"
      },
      "DocumentUpdate": {
        "properties": {
          "verification_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentVerificationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "verified_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified By"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "title": "DocumentUpdate"
      },
      "DocumentVerificationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "verified",
          "rejected",
          "reupload_requested"
        ],
        "title": "DocumentVerificationStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}